+++ src/frontend/ws-switcher-server.cc
@@ -392,7 +392,8 @@ public:
         ssl_session_.outbound_plaintext().push( s.bytes_written() );
       },
       [&] {
-        return ssl_session_.outbound_plaintext().writable_region().size() > ( 1 + WebSocketFrame::max_overhead() )
+        return ssl_session_.outbound_plaintext().writable_region().size()
+                 > size_t( 1 + WebSocketFrame::max_overhead() )
                and muxer_.output().readable_region().size() and ws_server_.handshake_complete();
       } ) );
